requirements.txt
opencv-python-headless
numpy
Pillow
matplotlib
scipy
tkinter
seaborn
albumentations
opencv-python
sklearn
torchvision
torch
xmltodict
tensorflow
albumentations.pytorch